06-agentcore-production-demo/cdk/requirements.txt
aws-cdk-lib>=2.244.0
constructs>=10.0.0
